pyproject.toml
[project]
name = "warpcore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=4.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.131.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.41.0",
]
